Move not used files for safekeeping
